Aggregate Depth Data Files

Function for combining high frequency data with depth.

agg_depth_files(filename_import, filename_export, dir_import, dir_export)

Arguments

filename_import

Vector of file names.

filename_export

File name for combined file.

dir_import

Directory for import data. Default is current working directory.

dir_export

Directory for export data. Default is current working directory.

Value

Returns a csv into the specified export directory of each file appended into a single file.

Details

Assumes files have the same column names. The files are merged with all=TRUE.

One new field (AggFile_Source) is added with the name of the source file. No other manipulation of the data is performed.

If any columns in the output have the ".x" or ".y" suffix when have duplicate column names in a file.

Examples

# Data Files
myFile_import <- c("Ellis--1.0m_Water_20180524_20180918.csv"
                  , "Ellis--3.0m_Water_20180524_20180918.csv")
myFile_export <- "Ellis--Combined_Water_20180524_20180918.csv"
myDir_import <- file.path(system.file("extdata", package = "LakeMonitoR"))
myDir_export <- tempdir()
agg_depth_files(filename_import = myFile_import
                , filename_export = myFile_export
                , dir_import = myDir_import
                , dir_export = myDir_export)
#> [1] "Total files to process = 2"
#> [1] "Processing item 1 of 2, COMPLETE, Ellis--1.0m_Water_20180524_20180918.csv."
#> [1] "Processing item 2 of 2, COMPLETE, Ellis--3.0m_Water_20180524_20180918.csv."