Lake Summary Statistics

Calculates summary statistics on daily depth means.

lake_summary_stats(data, col_date, col_depth, col_measure, below_threshold = 2)

Arguments

data
data frame

col_date
Column name, Date

col_depth
Column name, Depth

col_measure
Column name, measurement for calculation

below_threshold
Threshold below which to count number of days for measured value. Default = 2

Value

data frame in long format.

Details

Input data is assumed to be a single lake location depth profile with one measurement per day and depth. The function daily_depth_means can be used to generate this type of data from continuous data with multiple measurements per day for each depth.

Calculated statistics are:

* minimum

* maximum

* range

* mean

* median

* quantiles (1, 5, 10, 25, 50, 75, 90, 95, 99)

Statistics are calculated for multiple time periods:

* entire dataset

* year

* month

* year month

A data frame is returned in a hybrid long format. Statistics will be columns and time periods will also be a column with the values for each row. That is, if time period is month then each month will be a row with each statistic in a column.

Examples

# data
data <- laketemp_ddm

# Columns
col_date    <- "Date"
col_depth   <- "Depth"
col_measure <- "Measurement"
below_threshold <- 2

# Calculate Stratification
df_lss <- lake_summary_stats(data
                            , col_date
                            , col_depth
                            , col_measure
                            , below_threshold)

# Results
head(df_lss)

#>   TimeFrame_Name TimeFrame_Value   n ndays     mean   median       min      max
#> 1        AllData               2 357   357 8.496933 4.910854 0.3707292 22.08285
#> 2        AllData               3 357   357 8.493896 4.934021 0.4128542 22.04698
#> 3        AllData               4 357   357 8.412515 4.905479 0.4043333 21.87419
#> 4        AllData               5 357   357 8.320397 4.879625 0.3950208 21.81350
#> 5        AllData               6 357   357 8.196600 4.863479 0.3886667 20.88719
#> 6        AllData               7 357   357 8.018090 4.866729 0.4014583 19.93710
#>      range       sd      var        cv       q01       q05       q10      q25
#> 1 21.71213 7.455511 55.58465 0.8774356 0.4105175 0.5005958 0.6048125 1.406146
#> 2 21.63413 7.391324 54.63168 0.8701925 0.4527233 0.5380125 0.6436917 1.461646
#> 3 21.46985 7.333736 53.78368 0.8717649 0.4388183 0.5209042 0.6345750 1.435958
#> 4 21.41848 7.253241 52.60950 0.8717422 0.4255325 0.5311250 0.6309250 1.436063
#> 5 20.49852 7.117106 50.65319 0.8682997 0.4251250 0.5397125 0.6478000 1.446917
#> 6 19.53565 6.891263 47.48951 0.8594645 0.4382542 0.5636292 0.6885583 1.517688
#>        q50      q75      q90      q95      q99
#> 1 4.910854 16.36317 19.25685 19.86896 21.51376
#> 2 4.934021 16.29277 19.21007 19.82422 21.44097
#> 3 4.905479 16.20850 19.13412 19.75749 21.22852
#> 4 4.879625 16.04998 18.92680 19.70643 20.73644
#> 5 4.863479 15.89538 18.30049 19.61608 20.32941
#> 6 4.866729 15.66035 17.96345 18.96285 19.80878